Split components on entry pages
