+++ frontend/src/components/entity/EntityList.js
@@ -33,7 +33,7 @@ export default function EntityList({ entities }) {
   const classes = useStyles();
   const history = useHistory();
 
-  const keywordRef = useRef("");
+  const keywordRef = useRef({ value: "" });
   const [keyword, setKeyword] = useState("");
   const [page, setPage] = React.useState(0);
   const [rowsPerPage, setRowsPerPage] = React.useState(100);
