+++ frontend/src/components/entry/EntryForm.test.js
@@ -5,8 +5,8 @@
 import React from "react";
 import { render } from "@testing-library/react";
 
-import EntityForm from "./EntityForm";
+import EntryForm from "./EntryForm";
 
 test("should render a component with essential props", function () {
-  expect(() => render(<EntityForm />)).not.toThrow();
+  expect(() => render(<EntryForm entityId={1} />)).not.toThrow();
 });
