+++ frontend/src/components/entry/EntryList.test.js
@@ -5,8 +5,8 @@
 import React from "react";
 import { render } from "@testing-library/react";
 
-import EntityForm from "./EntityForm";
+import EntryList from "./EntryList";
 
 test("should render a component with essential props", function () {
-  expect(() => render(<EntityForm />)).not.toThrow();
+  expect(() => render(<EntryList entityId={0} entries={[]} />)).not.toThrow();
 });
